+++ src/components/shared-components/pop-up/one-btn-pop-up/OneBtnPopUp.tsx
@@ -1,20 +1,21 @@
 import PrimaryButton from "../../buttons/primary-button/PrimaryButton";
 import "../Popup.css"
 
-type InfoPopupProps = {
+type OneBtnPopUpProps = {
     heading: string,
     text: string,
-    okayAction: (close: boolean) => void
+    btnText: string,
+    onBtnClick: (close: boolean) => void
 }
 
-export default function InfoPopup({ heading, text, okayAction }: InfoPopupProps) {
+export default function OneBtnPopUp({ heading, text, onBtnClick, btnText }: OneBtnPopUpProps) {
     return (
         <div className="pop-up-overlay">
             <div className="pop-up-modal">
                 <h6 className="font-heading-h6" style={{ color: "#101828" }}>{heading}</h6>
                 <p className="font-md-regular" style={{ color: "#667085" }}> {text}</p>
                 <div className="pop-up-footer">
-                    <PrimaryButton label="Okay" size="small" onClick={() => okayAction(false)} />
+                    <PrimaryButton label={btnText} size="small" onClick={() => onBtnClick(false)} />
                 </div>
             </div>
         </div>
